+++ lib/grape-swagger.rb
@@ -150,7 +150,7 @@ module Grape
                               route.route_entity
                             end
 
-                  models = models.flatten.compact
+                  models = models_with_included_presenters(models.flatten.compact)
 
                   operation = {
                     notes: notes.to_s,
@@ -376,6 +376,26 @@ module Grape
               end
 
               result
+            end
+
+            def models_with_included_presenters(models)
+              all_models = models
+
+              models.each do |model|
+                properties = model.exposures
+                documented_properties = {}
+
+                model.documentation.keys.each_with_object(documented_properties) do |k, hash|
+                  hash[k] = properties[k] if properties.key?(k)
+                end
+
+                properties_configuration = documented_properties.values
+                additional_models = properties_configuration.map { |config| config[:using] }.compact
+
+                all_models += additional_models
+              end
+
+              all_models
             end
 
             def is_primitive?(type)