+++ spec/api_models_spec.rb
@@ -17,6 +17,28 @@ describe 'API Models' do
       end
     end
 
+    module Entities
+      class ComposedOf < Grape::Entity
+        expose :part_text, documentation: { type: 'string', desc: 'Content of composedof.' }
+      end
+
+      class ComposedOfElse < Grape::Entity
+        def self.entity_name
+          'composed'
+        end
+        expose :part_text, documentation: { type: 'string', desc: 'Content of composedof else.' }
+      end
+
+      class SomeThingElse < Grape::Entity
+        expose :else_text, documentation: { type: 'string', desc: 'Content of something else.' }
+        expose :parts, using: Entities::ComposedOf, documentation: { type: 'ComposedOf',
+                                                                     is_array: true,
+                                                                     required: true }
+
+        expose :part, using: Entities::ComposedOfElse, documentation: { type: 'composes' }
+      end
+    end
+
     class ModelsApi < Grape::API
       format :json
       desc 'This gets something.',
@@ -34,6 +56,15 @@ describe 'API Models' do
         thing = OpenStruct.new text: 'thing'
         present thing, with: Entities::Some::Thing
       end
+
+      desc 'This gets somthing else.',
+           entity: Entities::SomeThingElse
+      get '/somethingelse' do
+        part = OpenStruct.new part_text: 'part thing'
+        thing = OpenStruct.new else_text: 'else thing', parts: [part], part: part
+
+        present thing, with: Entities::SomeThingElse
+      end
       add_swagger_documentation
     end
   end
@@ -44,17 +75,20 @@ describe 'API Models' do
 
   it 'should document specified models' do
     get '/swagger_doc'
-    JSON.parse(last_response.body).should == {
-      'apiVersion' => '0.1',
-      'swaggerVersion' => '1.2',
-      'info' => {},
-      'produces' => ['application/json'],
-      'apis' => [
-        { 'path' => '/something.{format}', 'description' => 'Operations about somethings' },
-        { 'path' => '/thing.{format}', 'description' => 'Operations about things' },
-        { 'path' => '/swagger_doc.{format}', 'description' => 'Operations about swagger_docs' }
-      ]
-    }
+    result = JSON.parse(last_response.body)
+
+    expect(result).to include('apiVersion' => '0.1',
+                              'swaggerVersion' => '1.2',
+                              'info' => {},
+                              'produces' => ['application/json'])
+
+    expect(result['apis']).to eq [
+      { 'path' => '/something.{format}', 'description' => 'Operations about somethings' },
+      { 'path' => '/thing.{format}', 'description' => 'Operations about things' },
+      { 'path' => '/somethingelse.{format}',
+        'description' => 'Operations about somethingelses' },
+      { 'path' => '/swagger_doc.{format}', 'description' => 'Operations about swagger_docs' }
+    ]
   end
 
   it 'should include type when specified' do
@@ -121,6 +155,65 @@ describe 'API Models' do
         }
       }
     }
+  end
+
+  it 'should include entities which are only used as composition' do
+    get '/swagger_doc/somethingelse.json'
+    result = JSON.parse(last_response.body)
+
+    expect(result).to include('apiVersion' => '0.1',
+                              'swaggerVersion' => '1.2',
+                              'basePath' => 'http://example.org',
+                              'resourcePath' => '/somethingelse')
+
+    expect(result['apis']).to eq([{
+                                   'path' => '/somethingelse.{format}',
+                                   'operations' => [{
+                                     'notes' => '',
+                                     'type' => 'SomeThingElse',
+                                     'summary' => 'This gets somthing else.',
+                                     'nickname' => 'GET-somethingelse---format-',
+                                     'method' => 'GET',
+                                     'parameters' => []
+                                   }]
+                                 }])
+
+    expect(result['models']['SomeThingElse']).to include('id' => 'SomeThingElse',
+                                                         'properties' => {
+                                                           'else_text' => {
+                                                             'type' => 'string',
+                                                             'description' => 'Content of something else.'
+                                                           },
+                                                           'parts' => {
+                                                             'type' => 'array',
+                                                             'items' => { '$ref' => 'ComposedOf' }
+                                                           },
+                                                           'part' => { '$ref' => 'composes' }
+                                                         },
+                                                         'required' => ['parts']
+
+                                                 )
+
+    expect(result['models']['ComposedOf']).to include(
+                                                        'id' => 'ComposedOf',
+                                                        'properties' => {
+                                                          'part_text' => {
+                                                            'type' => 'string',
+                                                            'description' => 'Content of composedof.'
+                                                          }
+                                                        }
+                                                      )
+
+    expect(result['models']['composed']).to include(
+                                                      'id' => 'composed',
+                                                      'properties' => {
+                                                        'part_text' => {
+                                                          'type' => 'string',
+                                                          'description' => 'Content of composedof else.'
+                                                        }
+
+                                                      }
+                                                    )
   end
 
 end